Add ability to enable/disable task-switcher by creating/deleting a file
We (I) want to use the top-left button of the power board for
initiating the task switcher.  This is also used for robot start.
When one pressed it when the robot starts, one doesn't want to be
immediately brought into the task switcher.  So pyenv is going to
signal to sr-ts when it is OK for it to appear.

Why not just start sr-ts after the start button is pressed?  I don't
want to introduce this extra delay etc.